requirements.txt
django
djangorestframework
psycopg2
locust
flake8
drf-yasg[validation]
coverage
django-filter